requirements.txt
torchvision~=0.15.2
numpy==1.24.3
tqdm~=4.65.0
pandas~=1.5.3
pillow~=9.4.0
seaborn~=0.12.2
scikit-learn~=1.2.2
albumentations~=1.3.0